02. Classificador RA/requirements.txt
# python>=3.8
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.12.0
python-dotenv>=1.0.0
tqdm>=4.66.0
selenium>=4.15.0
webdriver-manager>=4.0.1
PyYAML>=6.0
